http client extension methods are used: PostAsJsonAsync, ReadFromJsonAsync
