+++ src/MinimalAPI/Program.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using MinimalAPI;
 using MinimalAPI.Auth;
 using MinimalAPI.DataSource.Tables;
+using MinimalAPI.Entities;
 using MinimalAPI.Extensions;
 using System.Security.Claims;
 
@@ -84,13 +85,3 @@ app.MapPost("/upload", (FileModel model) =>
 app.Run();
 
 public partial class Program { }
-
-internal record FileModel(IFormFile? File)
-{
-    public static async ValueTask<FileModel?> BindAsync(HttpContext context)
-    {
-        var form = await context.Request.ReadFormAsync();
-        var file = form.Files["file"];
-        return new FileModel(file);
-    }
-}
